# api/openapi-public.yaml
openapi: 3.0.3
info:
  title: Homeworldz Public API
  version: 0.1.0
  description: |
    The browser-facing API, consumed by the management site at
    `https://my.homeworldz.com` (see `web/src/config.js`) and by the Homeworldz
    client ([ADR 0030](../docs/adr/0030-client-architecture.md) and
    `docs/CLIENT2.md`). Served by the grid's `[website]` service, which listens
    on port 8003 by default.

    Distinct from `openapi.yaml`, which describes the internal grid API used by
    regions and tools rather than by browsers. The two share no credentials: a
    service token is not accepted here, and a website token is not accepted
    there.

    ## Conventions

    Every request body is one JSON object, at most 64 KiB, with **unknown
    fields rejected** — a misspelled property is `400 invalid_json` rather than
    a silently ignored field. Every error body is the `Error` schema, whose
    `code` is the stable machine-readable part; `message` is prose and may
    change.

    Two 404 codes mean different things and are deliberately distinguishable.
    `not_found` says the resource you named is absent or is not yours;
    `route_not_found` says this deployment serves no such endpoint, usually
    that it predates one. That way a client running ahead of a deployment can
    tell "not deployed" from "not allowed" without parsing prose.

    Requests may carry `X-Request-ID`; the value is echoed on the response, and
...

...
    `403 origin_forbidden`.

...
servers:
...
paths:
...
  /stats:
    get:
      operationId: getGridStats
      tags: [Discovery]
      summary: Read the public grid statistics
      description: |
        The figures published on the login page and the statistics page: how
        many accounts exist, how many people used one recently, how much land
        the grid runs and how much of it is up, and how long the grid has been
        running. Unauthenticated, because it is read by people who have no
        account yet.

        Every number is a count actually taken, never a placeholder. A figure
        that cannot be counted fails the whole request rather than appearing
        as a zero: a zero here reads as an exodus rather than as a database
        that was unreachable. Fields that have no answer yet — `uptimeSeconds`
        on a grid whose event log carries no start — are **absent** rather
        than zero, for the same reason.

        Readings are cached for up to 30 seconds, which is invisible against
        windows measured in days.

        The same figures are recorded once a day as a row of the grid's
        `stats.csv`, from the same code, so the page and the chart cannot
        disagree about what a column means.
      responses:
        "200":
          description: One reading of the grid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GridStats" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500":
          description: A figure could not be counted; nothing is published.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: |
            This deployment has no statistics sources configured, so it
            publishes nothing rather than zeros.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /registrations:
    post:
      operationId: registerAvatar
      tags: [Registration]
      summary: Register an avatar
      description: |
        Creates an unverified account and mails a confirmation code to the
        supplied address. The account cannot sign in — here or in a viewer —
        until `POST /verifications` consumes that code and sets a password.

        The `userid` is derived from the display name, which must be two words
        forming a 3-32 character `first.last` login ID. That derived userid is
        what login and password recovery take: the display name can change
        later, the userid cannot.

        A mail delivery failure does not fail the request. The account exists
        and the code can be re-requested from `POST /verifications/resend`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, email]
              additionalProperties: false
              properties:
                displayName:
                  type: string
                  description: Two words, forming a 3-32 character userid.
                  example: Jim Tarber
                email:
                  type: string
                  format: email
                  maxLength: 254
      responses:
        "201":
          description: The account was created and a confirmation code was mailed.
          headers:
            Location:
              description: Always `/v1/account`, readable once the caller holds a token.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [userid, displayName]
                properties:
                  userid: { type: string, example: jim.tarber }
                  displayName: { type: string, example: Jim Tarber }
        "400":
          description: "`invalid_display_name`, `invalid_email`, or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`userid_taken` — that display name is already registered."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /verifications:
    post:
      operationId: verifyRegistration
      tags: [Registration]
      summary: Confirm a registration and set its password
      description: |
        Consumes the emailed confirmation code, sets the account's password,
        and signs the caller in — the response is the same `TokenResponse` that
        `POST /tokens` returns, so a freshly verified account needs no second
        round trip.

        The code expires 24 hours after it was issued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, password]
              additionalProperties: false
              properties:
                code:
                  type: string
                  minLength: 1
                  maxLength: 128
                  description: The code from the confirmation email.
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
      responses:
        "200": { $ref: "#/components/responses/Token" }
        "400":
          description: "`invalid_code` (missing or oversized), `invalid_password`, or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            `invalid_code` — the code is unknown, already used, or expired. The
            status differs from the 400 above on purpose: this one says the
            request was well formed and the code was not accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /verifications/resend:
    post:
      operationId: resendVerification
      tags: [Registration]
      summary: Re-send a confirmation code
      description: |
        Mails a fresh confirmation code to the address on file.

        **The response is 202 whether or not the account exists**, whether or
        not it is already verified, and whether or not mail was sent, so this
        cannot be used to discover which accounts are registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userid]
              additionalProperties: false
              properties:
                userid:
                  type: string
                  description: The `first.last` login ID.
      responses:
        "202":
          description: Accepted. Returned for every outcome.
        "400":
          description: "`invalid_userid` (empty) or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /tokens:
    post:
      operationId: createToken
      tags: [Authentication]
      summary: Sign in and obtain a website token
      description: |
        Exchanges a userid and password for a short-lived website token. The
        token carries no in-world meaning: it authenticates this API only, and
        a viewer login is a separate exchange against the grid.

        A token stops being accepted the moment the account's authorization
        version changes, which a password change, a privilege change, and a ban
        all do. There is no refresh; sign in again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userid, password]
              additionalProperties: false
              properties:
                userid:
                  type: string
                  description: |
                    The `first.last` login ID. Case and spacing are normalized,
                    so `First Last` resolves to `first.last`.
                password:
                  type: string
                  format: password
                  maxLength: 128
      responses:
        "200": { $ref: "#/components/responses/Token" }
        "400":
          description: "`invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: |
            `invalid_credentials` — one reply for a wrong password, an unknown
            account, and an unverified one.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /password-resets:
    post:
      operationId: requestPasswordReset
      tags: [Authentication]
      summary: Request a password reset link
      description: |
        Sends a single-use reset link to the address on file for the named
        account. The link is only ever sent to that address, never to one
        supplied in the request.

        **The response is identical whether or not the account exists**, so this
        endpoint cannot be used to discover which accounts are registered. A
        caller learns only that the request was accepted.

        Rate limited. See [ADR 0034](../docs/adr/0034-password-recovery.md).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              additionalProperties: false
              properties:
                identifier:
                  type: string
                  minLength: 1
                  maxLength: 254
                  description: |
                    The login ID, in `first.last` form. This is the only accepted
                    identifier, and it is the same one login takes.

                    An email address is not accepted: nothing constrains it to a
                    single account, so it cannot name which password to change.
                    A display name is not accepted either — it can be changed,
                    while the login ID cannot. Case and spacing are normalized,
                    so `First Last` resolves to `first.last`.
      responses:
        "202":
          description: |
            Accepted. Returned whether or not an account matched, and whether or
            not mail was sent.
        "400":
          description: The request body was missing or malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429":
          description: Too many requests. The caller should retry later.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /password-resets/{token}:
    post:
      operationId: consumePasswordReset
      tags: [Authentication]
      summary: Set a new password using a reset token
      description: |
        Consumes a reset token and writes the new password. The token is
        single-use and expires one hour after it was issued.

        This performs exactly the same operation as an authenticated password
        change: it writes the password digests and nothing else. **Existing
        sessions are not ended** — a reset here means a forgotten password
        rather than a compromised account, and the reasoning is recorded in
        [ADR 0034](../docs/adr/0034-password-recovery.md).
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: The token from the emailed link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              additionalProperties: false
              properties:
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  format: password
      responses:
        "204":
          description: The password was changed. The caller may now sign in with it.
        "400":
          description: |
            The token is unknown, already used, or expired, or the new password
            fails the length rules. **These are deliberately one response**:
            distinguishing them would say whether a token had ever existed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429":
          description: Too many attempts. The caller should retry later.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /account:
    get:
      operationId: getAccount
      tags: [Account]
      summary: Read the signed-in account
      description: |
        Returns the caller's own identity, freshly loaded rather than decoded
        from the token, so a change made elsewhere is visible here immediately.
      security: [{ WebsiteToken: [] }]
      responses:
        "200":
          description: The caller's identity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /account/profile:
    patch:
      operationId: updateAccountProfile
      tags: [Account]
      summary: Change the signed-in account's display name
      description: |
        The display name is the only mutable part of an identity. The `userid`
        derived at registration is permanent, and changing the display name
        never changes it.
      security: [{ WebsiteToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName]
              additionalProperties: false
              properties:
                displayName: { type: string }
      responses:
        "200":
          description: The updated identity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "400":
          description: "`empty_update`, `invalid_display_name`, or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "409":
          description: "`display_name_taken`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /account/password:
    put:
      operationId: changeAccountPassword
      tags: [Account]
      summary: Change the signed-in account's password
      description: |
        Requires the current password even though the caller is already
        authenticated: the token proves the session, and this proves the person.

        A successful change bumps the account's authorization version, so
        **every existing token stops working, including the one that made this
        request**. The caller must sign in again.
      security: [{ WebsiteToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              additionalProperties: false
              properties:
                currentPassword: { type: string, format: password, maxLength: 128 }
                newPassword: { type: string, format: password, minLength: 8, maxLength: 128 }
      responses:
        "204":
          description: The password was changed.
        "400":
          description: "`wrong_password`, `invalid_password`, or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
...
  /client/inventory:
    get:
      operationId: listClientInventoryFolders
      tags: [Client]
      summary: List the caller's whole folder skeleton
      description: |
        Returns every folder the caller owns — the whole navigable skeleton in
        one call — plus the id of the root, so a client need not infer it from a
        parent that is empty and must not hard-code one.

        These are the same rows the viewer's AIS capability serves; only the
        credential and the encoding differ. **Read-only on purpose**: every
        mutation AIS supports carries ordering and permission rules that deserve
        their own design rather than a JSON transcription of the LLSD ones.
      security: [{ WebsiteToken: [] }]
      responses:
        "200":
          description: The root folder id and every folder the caller owns.
          content:
            application/json:
              schema:
                type: object
                required: [rootId, folders]
                properties:
                  rootId: { type: string, format: uuid }
                  folders:
                    type: array
                    items: { $ref: "#/components/schemas/InventoryFolder" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "503": { $ref: "#/components/responses/InventoryUnavailable" }
...
components:
...
  securitySchemes:
    WebsiteToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A website token from `POST /tokens` or `POST /verifications`. It is
        checked against the account's current authorization version on every
        request, so a password change, a privilege change, or a ban invalidates
        outstanding tokens at once. A region ticket from `POST /client/session`
        is not accepted here.
...